Cleanup LoggerStore
