+++ Sources/PulseCore/LoggerStore/LoggerStoreEvent.swift
