+++ Sources/PulseCore/LoggerStore/LoggerStoreLevel.swift
@@ -1,0 +1,48 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2020–2022 Alexander Grebenyuk (github.com/kean).
+
+extension LoggerStore {
+    public enum MetadataValue {
+        case string(String)
+        case stringConvertible(CustomStringConvertible)
+    }
+
+    public typealias Metadata = [String: MetadataValue]
+
+    // Compatible with SwiftLog.Logger.Level
+    @frozen public enum Level: String, CaseIterable, Codable, Hashable, Sendable {
+        case trace
+        case debug
+        case info
+        case notice
+        case warning
+        case error
+        case critical
+
+        var order: Int16 {
+            switch self {
+            case .trace: return 0
+            case .debug: return 1
+            case .info: return 2
+            case .notice: return 3
+            case .warning: return 4
+            case .error: return 5
+            case .critical: return 6
+            }
+        }
+    }
+}
+
+extension Dictionary where Key == String, Value == LoggerStore.MetadataValue {
+    func unpack() -> [String: String] {
+        var entries = [String: String]()
+        for (key, value) in self {
+            switch value {
+            case let .string(string): entries[key] = string
+            case let .stringConvertible(string): entries[key] = string.description
+            }
+        }
+        return entries
+    }
+}
